This is the fourth commit
In this commit, I added a function to validate whether the given number is a gapful number or not.

In this commit, I refactored the code a lot and learned new thing, used ArrayList but also used Queue in the middle.

I also used StingTokenizer, converting number to string, string to number etc.